img_io_3.patch : introduce signed / unsinged write + read methods
Patch based on img_io_2.patch by Ticker Berkin

git-svn-id: http://svn.mkgmap.org.uk/mkgmap/trunk@4167 25d90789-57f7-4ee0-8453-03a3dfeeeb22

+++ src/uk/me/parabola/imgfmt/app/map/Map.java
@@ -258,7 +258,7 @@ public class Map implements InternalFiles, Configurable {
 	 * @param flags The POI flags.
 	 */
 	public void addPoiDisplayFlags(int flags) {
-		treFile.addPoiDisplayFlags((byte) flags);
+		treFile.addPoiDisplayFlags(flags);
 	}
 
 	public void addMapObject(MapObject item) {
